requirements.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1602530275854/work
anytree==2.8.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==3.0.1
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1659291887007/work
audioread==3.0.0
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1625814693874/work
chardet @ file:///tmp/build/80754af9/chardet_1605303175790/work
charset-normalizer==2.0.12
clang==5.0
click==8.0.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1655412516417/work
cryptography @ file:///tmp/build/80754af9/cryptography_1616769282133/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958435052/work
decorator==4.4.2
Flask @ file:///opt/conda/conda-bld/flask_1648041541647/work
flatbuffers==1.12
gast==0.4.0
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1635085722655/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1623994295351/work
google-auth==1.35.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.48.2
gym==0.10.9
h5py==3.1.0
# Editable install with no version control (habitat==0.1.6)
habitat-sim==0.1.6
idna==3.4
idna-ssl @ file:///tmp/build/80754af9/idna_ssl_1611752490495/work
ifcfg==0.23
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1662975134926/work
imageio-ffmpeg @ file:///home/conda/feedstock_root/build_artifacts/imageio-ffmpeg_1649960641006/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1631916693255/work
importlib-resources==5.4.0
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
joblib==1.1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1610099771815/work
librosa==0.9.2
llvmlite==0.36.0
lmdb==1.3.0
magnum @ file:///opt/conda/envs/py36/conda-bld/habitat-sim_1600456622652/work/build/deps/magnum-bindings/src/python
Markdown @ file:///tmp/build/80754af9/markdown_1614363833670/work
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528150516/work
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-suite_1604870955430/work
multidict @ file:///tmp/build/80754af9/multidict_1600456401639/work
networkx==2.5.1
numba @ file:///home/conda/feedstock_root/build_artifacts/numba_1623568542230/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1626681920064/work
numpy-quaternion==2022.4.2
oauthlib @ file:///tmp/abs_08ngfezid4/croots/recipe/oauthlib_1659642459222/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
Pillow==8.4.0
pooch==1.6.0
protobuf==3.17.2
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyglet==1.5.26
PyJWT @ file:///tmp/build/80754af9/pyjwt_1619651623957/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305763431/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
PyWavelets==1.1.1
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.0
resampy==0.4.2
rigid-body-motion==0.9.1
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
smmap @ file:///home/conda/feedstock_root/build_artifacts/smmap_1611376390914/work
soundfile==0.11.0
tensorboard @ file:///tmp/build/80754af9/tensorboard_1633093581375/work/tensorboard-2.6.0-py3-none-any.whl
tensorboard-data-server @ file:///tmp/build/80754af9/tensorboard-data-server_1633035063888/work/tensorboard_data_server-0.6.0-py3-none-manylinux2010_x86_64.whl
tensorboard-plugin-wit==1.6.0
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.10.0
torchvision==0.2.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094701020/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1662214488106/work
typing-extensions==3.7.4.3
urllib3==1.26.12
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
wrapt==1.12.1
yacs==0.1.8
yarl @ file:///tmp/build/80754af9/yarl_1606939915466/work
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
